LRNT-004: Creating portfolio ECR/ECS template (#6)
## 🧑‍💻 Description
These changes enable the creation of the module for the curriculum web
application (back-end and front-end). They also disable temporarily the
load balancer and DNS records specifications. So, this will only create
the ECR empty repository and the ECS empty cluster.

## ✅ Testing
The changes were merged to `development` and then to `staging` branches
to test the terraform pipeline with them.

## 🖼️ Evidence
Confirming the repository was created in AWS development account:

![image](https://github.com/zatarain/lorentz/assets/539783/c60709be-cc44-47e1-a428-4cec4f32e466)

And also in AWS staging account:

![image](https://github.com/zatarain/lorentz/assets/539783/3e48d132-d2cd-4259-8719-7cf5f7e7c5e1)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>