LRNT-009: Unifying security groups for the application (#15)
## 🧑‍💻 Description
These changes create a new security group for the application in order
to be used by front-end and back-end. The purpose is use less AWS
resources if necessary. After deploy these changes, we should remove the
older security groups to keep only the new ones.

## ✅ Testing
These changes were deployed to `development` and `staging` environments
and the services still working there after the deployment.

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>